+++ Sources/BaseFoundationiOS/Common/Logger/OSLogger.swift
@@ -1,0 +1,13 @@
+//
+//  OSLogger.swift
+//
+//
+//  Created by pablo borquez on 10/07/23.
+//
+
+import Foundation
+import OSLog
+
+open class OSLogger: CoreLogger {
+    
+}